refactor: move all frontend pages into frontend/pages directory for proper Vite resolution
