+++ frontend/src/main.jsx
@@ -18,8 +18,8 @@ import WelcomeOverlay from './WelcomeOverlay.jsx';
 import WelcomeExperience from './WelcomeExperience.jsx';
 import OnboardingOverlay from './OnboardingOverlay.jsx';
 import Gallery from './Gallery.jsx';
-import AgentsPage from '../../pages/Agents.jsx';
-import Dashboard from '../../pages/Dashboard.jsx';
+import AgentsPage from '../pages/Agents.jsx';
+import Dashboard from '../pages/Dashboard.jsx';
 import FlowViewPage from '../pages/FlowViewPage.jsx';
 import Sandbox from './Sandbox.jsx';
 import FeedbackFab from './FeedbackFab.jsx';
